requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
scikit-learn
joblib
passlib[bcrypt]
PyJWT
python-jose[cryptography]
sqlalchemy
psycopg2-binary
boto3
supabase
python-dotenv